+++ maps-sdk/src/main/java/com/mapbox/maps/coroutine/MapboxMapExt.kt
@@ -6,7 +6,6 @@
 
 package com.mapbox.maps.coroutine
 
-import androidx.annotation.RestrictTo
 import com.mapbox.bindgen.Expected
 import com.mapbox.bindgen.None
 import com.mapbox.bindgen.Value
@@ -39,7 +38,6 @@ suspend fun MapboxMap.awaitStyle(): Style = suspendCoroutine { continuation ->
  * Load a new style from a style extension, suspends until style is loaded.
  *
  * @param styleExtension the style extension to load
- * @param transitionOptions the transition options to use when loading the style
  */
 @JvmSynthetic
 suspend fun MapboxMap.awaitLoadStyle(
@@ -54,8 +52,7 @@ suspend fun MapboxMap.awaitLoadStyle(
 /**
  * Load a new style from a style URI or JSON, suspends until style is loaded.
  *
- * @param styleUri the style URI to load
- * @param transitionOptions the transition options to use when loading the style
+ * @param style the style URI to load
  */
 @JvmSynthetic
 suspend fun MapboxMap.awaitLoadStyle(
@@ -312,7 +309,6 @@ val MapboxMap.cameraChangedEvents: Flow<CameraChanged>
 /**
  * Conflated [Flow] of [CameraChangedCoalesced] updates from [MapboxMap.subscribeCameraChangedCoalesced].
  */
-@get:RestrictTo(RestrictTo.Scope.LIBRARY_GROUP_PREFIX)
 @com.mapbox.annotation.MapboxExperimental
 val MapboxMap.cameraChangedCoalescedEvents: Flow<CameraChangedCoalesced>
   @JvmSynthetic